backend/requirements.txt
# I-Fridge Backend — Python Dependencies
# ========================================

# --- Core Framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.9

# --- Database (Supabase) ---
supabase==2.9.0
asyncpg==0.30.0

# --- AI / Vision ---
clarifai-grpc==10.8.0

# --- Computation ---
numpy==2.1.0

# --- Utilities ---
python-dotenv==1.0.1
httpx==0.27.0

# --- Testing ---
pytest==8.3.0
pytest-asyncio==0.24.0
